fix(exploration): key syscall stats by i128 so u64::MAX can't alias the -1 sentinel
handle_syscall_core's three stats sites keyed HashMap<i64, u64> by
num.map(|n| n as i64).unwrap_or(-1), where -1 means "the syscall register
was symbolic at dispatch". u64::MAX as i64 is also -1, so a guest-controlled
register concretized to 0xFFFF_FFFF_FFFF_FFFF had its count misattributed to
the symbolic bucket. Widening the key to i128 puts every representable u64 in
the non-negative half, leaving the sentinel unreachable; the Python-visible
dict keys are unchanged for every realistic syscall number.

angr-0jh0j.18

+++ native/angr/src/exploration/core_outcome.rs
@@ -221,9 +221,15 @@ pub(crate) struct CoreCounters {
     pub(crate) not_implemented_fallbacks_by_name: HashMap<String, u64>,
     pub(crate) other_fallbacks_by_name: HashMap<String, u64>,
     pub(crate) syscall_native_count: u64,
-    pub(crate) syscall_native_by_num: HashMap<i64, u64>,
+    /// Keyed by `num as i128`, not `as i64`: every real `u64` syscall number
+    /// then lands in the non-negative half, keeping the `-1` "syscall register
+    /// was symbolic" sentinel unreachable. See the field of the same name on
+    /// [`RustExplorationManager`](crate::exploration::RustExplorationManager).
+    pub(crate) syscall_native_by_num: HashMap<i128, u64>,
     pub(crate) syscall_python_fallback_count: u64,
-    pub(crate) syscall_python_fallback_by_num: HashMap<i64, u64>,
+    /// Keyed by `num as i128` for the same sentinel-disjointness reason as
+    /// `syscall_native_by_num` above.
+    pub(crate) syscall_python_fallback_by_num: HashMap<i128, u64>,
     pub(crate) simprocedure_python_fallback_count: u64,
     pub(crate) simprocedure_fallback_by_name: HashMap<String, u64>,
     pub(crate) deferred_forks_dropped: u64,
